requirements.txt
absl-py==2.3.1
accelerate==1.1.0
addict==2.4.0
affine==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
aniso8601==10.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
anywidget==0.9.18
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
blinker==1.9.0
bqplot==0.12.44
branca==0.8.1
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
color-operations==0.2.0
colorama==0.4.6
colour==0.1.5
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.6.0
decorator==5.2.1
diffusers==0.32.2
dill==0.3.8
distlib==0.3.9
duckdb==1.2.2
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
facexlib==0.3.0
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
filterpy==1.4.5
fiona==1.10.1
Flask==3.1.0
Flask-Caching==2.3.1
flask-cors==5.0.1
flask-restx==1.3.0
folium==0.19.5
fonttools==4.56.0
frozenlist==1.7.0
fsspec==2025.2.0
ftfy==6.3.1
future==1.0.0
gdown==5.2.0
geojson==3.2.0
geopandas==1.0.1
gradio==5.18.0
gradio_client==1.7.2
groundingdino-py==0.4.0
grpcio==1.73.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
hydra-core==1.3.2
icecream==2.1.5
identify==2.6.12
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
iopath==0.1.10
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.19.2
ipympl==0.9.7
ipython==8.36.0
ipytree==0.2.2
ipyvue==1.11.2
ipyvuetify==1.11.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-leaflet==0.19.2
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
leafmap==0.43.13
llvmlite==0.44.0
lmdb==1.6.2
localtileserver==0.10.6
lpips==0.1.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
morecantile==6.2.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
narwhals==1.38.1
networkx==3.4.2
nodeenv==1.9.1
numba==0.61.2
numexpr==2.10.2
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
patool==4.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.8
plotly==6.0.1
pluggy==1.6.0
portalocker==3.1.1
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psygnal==0.13.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycocotools==2.0.8
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyiqa==0.1.14.1
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
pyshp==2.3.1
PySocks==1.7.1
pystac==1.13.0
pystac-client==0.8.6
pytest==8.4.1
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
rasterio==1.4.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rio-cogeo==5.4.1
rio-tiler==7.7.0
rioxarray==0.19.0
rpds-py==0.24.0
ruff==0.9.7
safehttpx==0.1.6
safetensors==0.5.3
sam2==1.1.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
scooby==0.10.1
seaborn==0.13.2
segment-anything==1.0
segment-anything-hq==0.3
segment-anything-py==1.0.1
segment-geospatial==0.12.6
semantic-version==2.10.0
sentencepiece==0.2.0
server-thread==0.3.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.45.3
supervision==0.6.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tokenizers==0.15.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
transformers==4.37.2
triton==3.2.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.99
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.31.2
wcwidth==0.2.13
websockets==15.0
Werkzeug==3.1.3
whitebox==2.3.6
whiteboxgui==2.3.0
widgetsnbextension==4.0.14
xarray==2025.4.0
xxhash==3.5.0
xyzservices==2025.4.0
yapf==0.43.0
yarl==1.20.1
zipp==3.21.0
